pyproject.toml
[project]
name = "ejpipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "dynaconf>=3.2.6",
    "htmllistparse>=0.6.1",
    "pytest-playwright>=0.6.2",
    "requests>=2.32.3",
    "rich>=13.9.4",
]

[dependency-groups]
dev = [
    "coverage>=7.6.8",
    "ipython>=8.30.0",
    "pytest>=8.3.4",
    "ruff>=0.8.2",
]
